sync interface method implementation

+++ Interfaces/IGarbagePoweredOptions.cs
@@ -3,7 +3,7 @@
 public interface IGarbagePoweredOptions
 
 {
-    public int ProductionPerUnit { get; set; }
+    public float ProductionPerUnit { get; set; }
     public int ProductionCapacity { get; set; }
 
 }
