+++ Interfaces/IParkingFacilityOptions.cs
@@ -4,5 +4,5 @@ public interface IParkingFacilityOptions
 
 {
     public int GarageCapacity { get; set; }
-    public int Comfortfactor { get; set; }
+    public float Comfortfactor { get; set; }
 }
