+++ Interfaces/ISewageOutletOptions.cs
@@ -3,6 +3,6 @@
 public interface ISewageOutletOptions
 {
     public int SewageCapacity { get; set; }
-    public int SewagePurification { get; set; }
+    public float SewagePurification { get; set; }
 
 }
